wip(#1102): extract emit/js_fallback.rs + emit/string_collection.rs
