+++ crates/perry-codegen-wasm/src/emit/mod.rs
@@ -9,10 +9,12 @@ mod binary;
 mod closures;
 mod expr;
 mod function;
+mod js_fallback;
 mod locals;
 mod memcall;
 mod method_call;
 mod stmt;
+mod string_collection;
 
 use closures::{collect_closures_from_expr, collect_closures_from_stmts};
 use locals::{collect_locals, collect_module_let_ids, resolve_source_module_idx};
@@ -2182,1594 +2184,6 @@ impl WasmModuleEmitter {
             wasm_bytes,
             async_js,
             ffi_imports: ffi_import_names,
-        }
-    }
-
-    /// Generate JavaScript code for an async function body.
-    /// The generated function uses NaN-boxed f64 values and bridge helper functions.
-    fn emit_js_async_function(&self, func: &perry_hir::ir::Function) -> String {
-        let params: Vec<String> = func
-            .params
-            .iter()
-            .enumerate()
-            .map(|(i, _)| format!("__p{}", i))
-            .collect();
-        let params_str = params.join(", ");
-
-        let mut body = String::new();
-        // Map param names to local IDs for the JS emitter
-        let mut local_names: BTreeMap<u32, String> = BTreeMap::new();
-        for (i, param) in func.params.iter().enumerate() {
-            local_names.insert(param.id, format!("__p{}", i));
-        }
-
-        for stmt in &func.body {
-            self.emit_js_stmt(&mut body, stmt, &mut local_names, 2);
-        }
-
-        format!(
-            "  __async_{name}: ({params}) => {{\n\
-             \x20   const __p = (async () => {{\n\
-             {body}\
-             \x20     return {undef};\n\
-             \x20   }})();\n\
-             \x20   return nanboxPointer(allocHandle(__p));\n\
-             \x20 }},",
-            name = func.name,
-            params = params_str,
-            body = body,
-            undef = "u64ToF64(TAG_UNDEFINED)",
-        )
-    }
-
-    fn emit_js_stmt(
-        &self,
-        out: &mut String,
-        stmt: &Stmt,
-        locals: &mut BTreeMap<u32, String>,
-        indent: usize,
-    ) {
-        let pad = "  ".repeat(indent);
-        match stmt {
-            Stmt::Let { id, init, .. } => {
-                let name = format!("__l{}", id);
-                locals.insert(*id, name.clone());
-                if let Some(init_expr) = init {
-                    let val = self.emit_js_expr(init_expr, locals);
-                    out.push_str(&format!("{pad}    let {name} = {val};\n"));
-                } else {
-                    out.push_str(&format!("{pad}    let {name} = u64ToF64(TAG_UNDEFINED);\n"));
-                }
-            }
-            Stmt::Expr(e) => {
-                let val = self.emit_js_expr(e, locals);
-                out.push_str(&format!("{pad}    {val};\n"));
-            }
-            Stmt::Return(Some(e)) => {
-                let val = self.emit_js_expr(e, locals);
-                out.push_str(&format!("{pad}    return {val};\n"));
-            }
-            Stmt::Return(None) => {
-                out.push_str(&format!("{pad}    return u64ToF64(TAG_UNDEFINED);\n"));
-            }
-            Stmt::If {
-                condition,
-                then_branch,
-                else_branch,
-            } => {
-                let cond = self.emit_js_expr(condition, locals);
-                out.push_str(&format!("{pad}    if (toJsValue({cond})) {{\n"));
-                for s in then_branch {
-                    self.emit_js_stmt(out, s, locals, indent + 1);
-                }
-                if let Some(eb) = else_branch {
-                    out.push_str(&format!("{pad}    }} else {{\n"));
-                    for s in eb {
-                        self.emit_js_stmt(out, s, locals, indent + 1);
-                    }
-                }
-                out.push_str(&format!("{pad}    }}\n"));
-            }
-            Stmt::While { condition, body } => {
-                let cond = self.emit_js_expr(condition, locals);
-                out.push_str(&format!("{pad}    while (toJsValue({cond})) {{\n"));
-                for s in body {
-                    self.emit_js_stmt(out, s, locals, indent + 1);
-                }
-                out.push_str(&format!("{pad}    }}\n"));
-            }
-            Stmt::For {
-                init,
-                condition,
-                update,
-                body,
-            } => {
-                out.push_str(&format!("{pad}    {{\n"));
-                if let Some(init_stmt) = init {
-                    self.emit_js_stmt(out, init_stmt, locals, indent + 1);
-                }
-                let cond = condition
-                    .as_ref()
-                    .map(|c| self.emit_js_expr(c, locals))
-                    .unwrap_or_else(|| "1".to_string());
-                out.push_str(&format!("{pad}      while (toJsValue({cond})) {{\n"));
-                for s in body {
-                    self.emit_js_stmt(out, s, locals, indent + 2);
-                }
-                if let Some(upd) = update {
-                    let u = self.emit_js_expr(upd, locals);
-                    out.push_str(&format!("{pad}        {u};\n"));
-                }
-                out.push_str(&format!("{pad}      }}\n"));
-                out.push_str(&format!("{pad}    }}\n"));
-            }
-            Stmt::Try {
-                body,
-                catch,
-                finally,
-            } => {
-                out.push_str(&format!("{pad}    try {{\n"));
-                for s in body {
-                    self.emit_js_stmt(out, s, locals, indent + 1);
-                }
-                if let Some(c) = catch {
-                    if let Some((param_id, _)) = &c.param {
-                        let name = format!("__l{}", param_id);
-                        locals.insert(*param_id, name.clone());
-                        out.push_str(&format!("{pad}    }} catch (__e) {{\n"));
-                        out.push_str(&format!("{pad}      let {name} = fromJsValue(__e);\n"));
-                    } else {
-                        out.push_str(&format!("{pad}    }} catch (__e) {{\n"));
-                    }
-                    for s in &c.body {
-                        self.emit_js_stmt(out, s, locals, indent + 1);
-                    }
-                }
-                if let Some(f) = finally {
-                    out.push_str(&format!("{pad}    }} finally {{\n"));
-                    for s in f {
-                        self.emit_js_stmt(out, s, locals, indent + 1);
-                    }
-                }
-                out.push_str(&format!("{pad}    }}\n"));
-            }
-            Stmt::Throw(e) => {
-                let val = self.emit_js_expr(e, locals);
-                out.push_str(&format!("{pad}    throw toJsValue({val});\n"));
-            }
-            Stmt::Break => {
-                out.push_str(&format!("{pad}    break;\n"));
-            }
-            Stmt::Continue => {
-                out.push_str(&format!("{pad}    continue;\n"));
-            }
-            Stmt::LabeledBreak(label) => {
-                out.push_str(&format!("{pad}    break {};\n", label));
-            }
-            Stmt::LabeledContinue(label) => {
-                out.push_str(&format!("{pad}    continue {};\n", label));
-            }
-            Stmt::DoWhile { body, condition } => {
-                out.push_str(&format!("{pad}    do {{\n"));
-                for s in body {
-                    self.emit_js_stmt(out, s, locals, indent + 1);
-                }
-                let cond = self.emit_js_expr(condition, locals);
-                out.push_str(&format!("{pad}    }} while (isTruthy({cond}));\n"));
-            }
-            Stmt::Labeled { label, body } => {
-                out.push_str(&format!("{pad}    {}: {{\n", label));
-                self.emit_js_stmt(out, body, locals, indent + 1);
-                out.push_str(&format!("{pad}    }}\n"));
-            }
-            Stmt::Switch {
-                discriminant,
-                cases,
-            } => {
-                let disc = self.emit_js_expr(discriminant, locals);
-                out.push_str(&format!("{pad}    switch (toJsValue({disc})) {{\n"));
-                for case in cases {
-                    if let Some(test) = &case.test {
-                        let t = self.emit_js_expr(test, locals);
-                        out.push_str(&format!("{pad}      case toJsValue({t}):\n"));
-                    } else {
-                        out.push_str(&format!("{pad}      default:\n"));
-                    }
-                    for s in &case.body {
-                        self.emit_js_stmt(out, s, locals, indent + 2);
-                    }
-                }
-                out.push_str(&format!("{pad}    }}\n"));
-            }
-            // Issue #569: PreallocateBoxes is a perry-codegen-only directive
-            // — JS hoisting handles forward refs natively, so the wasm/JS
-            // backend has no equivalent to emit.
-            Stmt::PreallocateBoxes(_) => {}
-        }
-    }
-
-    fn emit_js_expr(&self, expr: &Expr, locals: &BTreeMap<u32, String>) -> String {
-        match expr {
-            Expr::Number(n) => format!("{}", n),
-            Expr::Integer(i) => format!("{}", *i as f64),
-            Expr::Bool(true) => "u64ToF64(TAG_TRUE)".to_string(),
-            Expr::Bool(false) => "u64ToF64(TAG_FALSE)".to_string(),
-            Expr::Undefined => "u64ToF64(TAG_UNDEFINED)".to_string(),
-            Expr::Null => "u64ToF64(TAG_NULL)".to_string(),
-            Expr::String(s) => {
-                let escaped = s
-                    .replace('\\', "\\\\")
-                    .replace('"', "\\\"")
-                    .replace('\n', "\\n")
-                    .replace('\r', "\\r");
-                format!("fromJsValue(\"{}\")", escaped)
-            }
-            Expr::LocalGet(id) => locals
-                .get(id)
-                .cloned()
-                .unwrap_or_else(|| format!("__l{}", id)),
-            Expr::LocalSet(id, val) => {
-                let name = locals
-                    .get(id)
-                    .cloned()
-                    .unwrap_or_else(|| format!("__l{}", id));
-                let v = self.emit_js_expr(val, locals);
-                format!("({} = {})", name, v)
-            }
-            Expr::GlobalGet(id) => format!("__g{}", id),
-            Expr::GlobalSet(id, val) => {
-                let v = self.emit_js_expr(val, locals);
-                format!("(__g{} = {})", id, v)
-            }
-            Expr::Binary { op, left, right } => {
-                let l = self.emit_js_expr(left, locals);
-                let r = self.emit_js_expr(right, locals);
-                match op {
-                    BinaryOp::Add => format!("fromJsValue(toJsValue({}) + toJsValue({}))", l, r),
-                    BinaryOp::Sub => format!("({} - {})", l, r),
-                    BinaryOp::Mul => format!("({} * {})", l, r),
-                    BinaryOp::Div => format!("({} / {})", l, r),
-                    BinaryOp::Mod => format!("({} % {})", l, r),
-                    _ => format!("fromJsValue(toJsValue({}) + toJsValue({}))", l, r),
-                }
-            }
-            Expr::Compare { op, left, right } => {
-                let l = self.emit_js_expr(left, locals);
-                let r = self.emit_js_expr(right, locals);
-                let js_op = match op {
-                    CompareOp::Eq => "===",
-                    CompareOp::Ne => "!==",
-                    CompareOp::LooseEq => "==",
-                    CompareOp::LooseNe => "!=",
-                    CompareOp::Lt => "<",
-                    CompareOp::Le => "<=",
-                    CompareOp::Gt => ">",
-                    CompareOp::Ge => ">=",
-                };
-                format!(
-                    "(toJsValue({}) {} toJsValue({}) ? u64ToF64(TAG_TRUE) : u64ToF64(TAG_FALSE))",
-                    l, js_op, r
-                )
-            }
-            Expr::Logical { op, left, right } => {
-                let l = self.emit_js_expr(left, locals);
-                let r = self.emit_js_expr(right, locals);
-                match op {
-                    LogicalOp::And => format!("(toJsValue({l}) ? {r} : {l})"),
-                    LogicalOp::Or => format!("(toJsValue({l}) ? {l} : {r})"),
-                    LogicalOp::Coalesce => format!("(isNull({l}) || isUndefined({l}) ? {r} : {l})"),
-                }
-            }
-            Expr::Unary { op, operand } => {
-                let o = self.emit_js_expr(operand, locals);
-                match op {
-                    UnaryOp::Neg => format!("(-{})", o),
-                    UnaryOp::Not => format!(
-                        "(toJsValue({}) ? u64ToF64(TAG_FALSE) : u64ToF64(TAG_TRUE))",
-                        o
-                    ),
-                    _ => o,
-                }
-            }
-            Expr::Await(inner) => {
-                let v = self.emit_js_expr(inner, locals);
-                // In JS async context, we can truly await
-                format!("fromJsValue(await toJsValue({}))", v)
-            }
-            Expr::Call { callee, args, .. } => {
-                let args_js: Vec<String> =
-                    args.iter().map(|a| self.emit_js_expr(a, locals)).collect();
-                match callee.as_ref() {
-                    Expr::FuncRef(id) => {
-                        if let Some(&func_idx) = self.func_map.get(id) {
-                            // Call exported WASM function. WASM funcs use i64 params/result.
-                            let args_i64: Vec<String> =
-                                args_js.iter().map(|a| format!("f64ToU64({})", a)).collect();
-                            format!(
-                                "u64ToF64(wasmInstance.exports.__wasm_func_{}({}))",
-                                func_idx,
-                                args_i64.join(", ")
-                            )
-                        } else {
-                            "u64ToF64(TAG_UNDEFINED)".to_string()
-                        }
-                    }
-                    Expr::ExternFuncRef { name, .. } => {
-                        if let Some(&func_idx) = self.func_name_map.get(name) {
-                            let args_i64: Vec<String> =
-                                args_js.iter().map(|a| format!("f64ToU64({})", a)).collect();
-                            format!(
-                                "u64ToF64(wasmInstance.exports.__wasm_func_{}({}))",
-                                func_idx,
-                                args_i64.join(", ")
-                            )
-                        } else {
-                            "u64ToF64(TAG_UNDEFINED)".to_string()
-                        }
-                    }
-                    Expr::PropertyGet { object, property } => {
-                        let obj = self.emit_js_expr(object, locals);
-                        let _args_str = args_js.join(", ");
-                        format!(
-                            "fromJsValue(toJsValue({}).{}({}))",
-                            obj,
-                            property,
-                            args.iter()
-                                .map(|a| format!("toJsValue({})", self.emit_js_expr(a, locals)))
-                                .collect::<Vec<_>>()
-                                .join(", ")
-                        )
-                    }
-                    _ => {
-                        let callee_js = self.emit_js_expr(callee, locals);
-                        format!("{}({})", callee_js, args_js.join(", "))
-                    }
-                }
-            }
-            Expr::FetchWithOptions {
-                url,
-                method,
-                body,
-                headers,
-            } => {
-                let url_js = self.emit_js_expr(url, locals);
-                let method_js = self.emit_js_expr(method, locals);
-                let body_js = self.emit_js_expr(body, locals);
-                // In async JS context, we can do a real fetch
-                let mut opts = format!("{{ method: getString({}) || 'GET'", method_js);
-                if !matches!(body.as_ref(), Expr::Undefined) {
-                    opts.push_str(&format!(", body: getString({})", body_js));
-                }
-                if !headers.is_empty() {
-                    opts.push_str(", headers: {");
-                    for (i, (key, val)) in headers.iter().enumerate() {
-                        if i > 0 {
-                            opts.push_str(", ");
-                        }
-                        let v = self.emit_js_expr(val, locals);
-                        opts.push_str(&format!("'{}': getString({})", key, v));
-                    }
-                    opts.push('}');
-                }
-                opts.push('}');
-                format!("fromJsValue(await fetch(getString({}), {}))", url_js, opts)
-            }
-            Expr::PropertyGet { object, property } => {
-                let obj = self.emit_js_expr(object, locals);
-                format!("fromJsValue(toJsValue({}).{})", obj, property)
-            }
-            Expr::PropertySet {
-                object,
-                property,
-                value,
-            } => {
-                let obj = self.emit_js_expr(object, locals);
-                let val = self.emit_js_expr(value, locals);
-                format!(
-                    "(toJsValue({}).{} = toJsValue({}), {})",
-                    obj, property, val, val
-                )
-            }
-            Expr::Object(fields) => {
-                let mut parts = Vec::new();
-                for (key, val) in fields {
-                    let v = self.emit_js_expr(val, locals);
-                    parts.push(format!("'{}': toJsValue({})", key, v));
-                }
-                format!("fromJsValue({{{}}})", parts.join(", "))
-            }
-            Expr::Array(elements) => {
-                let elems: Vec<String> = elements
-                    .iter()
-                    .map(|e| format!("toJsValue({})", self.emit_js_expr(e, locals)))
-                    .collect();
-                format!("fromJsValue([{}])", elems.join(", "))
-            }
-            Expr::Conditional {
-                condition,
-                then_expr,
-                else_expr,
-            } => {
-                let c = self.emit_js_expr(condition, locals);
-                let t = self.emit_js_expr(then_expr, locals);
-                let e = self.emit_js_expr(else_expr, locals);
-                format!("(toJsValue({}) ? {} : {})", c, t, e)
-            }
-            Expr::NativeMethodCall {
-                module,
-                method,
-                object,
-                args,
-                class_name,
-            } => {
-                let normalized = module.strip_prefix("node:").unwrap_or(module);
-                match normalized {
-                    "console" => {
-                        let args_js: Vec<String> = args
-                            .iter()
-                            .map(|a| format!("toJsValue({})", self.emit_js_expr(a, locals)))
-                            .collect();
-                        match method.as_str() {
-                            "log" => format!(
-                                "(console.log({}), u64ToF64(TAG_UNDEFINED))",
-                                args_js.join(", ")
-                            ),
-                            "warn" => format!(
-                                "(console.warn({}), u64ToF64(TAG_UNDEFINED))",
-                                args_js.join(", ")
-                            ),
-                            "error" => format!(
-                                "(console.error({}), u64ToF64(TAG_UNDEFINED))",
-                                args_js.join(", ")
-                            ),
-                            _ => "u64ToF64(TAG_UNDEFINED)".to_string(),
-                        }
-                    }
-                    "JSON" => match method.as_str() {
-                        "parse" if !args.is_empty() => {
-                            let a = self.emit_js_expr(&args[0], locals);
-                            format!("fromJsValue(JSON.parse(getString({})))", a)
-                        }
-                        "stringify" if !args.is_empty() => {
-                            let a = self.emit_js_expr(&args[0], locals);
-                            format!("fromJsValue(JSON.stringify(toJsValue({})))", a)
-                        }
-                        _ => "u64ToF64(TAG_UNDEFINED)".to_string(),
-                    },
-                    "perry/ui" | "perry/system" => {
-                        let bridge_name = map_ui_method(method, class_name.as_deref());
-                        let args_js: Vec<String> =
-                            args.iter().map(|a| self.emit_js_expr(a, locals)).collect();
-                        if let Some(obj) = object {
-                            let obj_js = self.emit_js_expr(obj, locals);
-                            let mut all_args = vec![obj_js];
-                            all_args.extend(args_js);
-                            format!("__perryUi.{}({})", bridge_name, all_args.join(", "))
-                        } else {
-                            format!("__perryUi.{}({})", bridge_name, args_js.join(", "))
-                        }
-                    }
-                    _ => {
-                        if let Some(obj) = object {
-                            let obj_js = self.emit_js_expr(obj, locals);
-                            let args_js: Vec<String> = args
-                                .iter()
-                                .map(|a| format!("toJsValue({})", self.emit_js_expr(a, locals)))
-                                .collect();
-                            format!(
-                                "fromJsValue(toJsValue({}).{}({}))",
-                                obj_js,
-                                method,
-                                args_js.join(", ")
-                            )
-                        } else {
-                            "u64ToF64(TAG_UNDEFINED)".to_string()
-                        }
-                    }
-                }
-            }
-            Expr::ErrorNew(msg) => {
-                if let Some(m) = msg {
-                    let m_js = self.emit_js_expr(m, locals);
-                    format!("fromJsValue(new Error(getString({})))", m_js)
-                } else {
-                    "fromJsValue(new Error())".to_string()
-                }
-            }
-            Expr::ErrorMessage(err) => {
-                let e = self.emit_js_expr(err, locals);
-                format!("fromJsValue(toJsValue({}).message)", e)
-            }
-            Expr::ErrorNewWithCause { message, cause } => {
-                let m_js = self.emit_js_expr(message, locals);
-                let c_js = self.emit_js_expr(cause, locals);
-                format!(
-                    "fromJsValue(new Error(getString({}), {{ cause: toJsValue({}) }}))",
-                    m_js, c_js
-                )
-            }
-            Expr::TypeErrorNew(m) => {
-                let m_js = self.emit_js_expr(m, locals);
-                format!("fromJsValue(new TypeError(getString({})))", m_js)
-            }
-            Expr::RangeErrorNew(m) => {
-                let m_js = self.emit_js_expr(m, locals);
-                format!("fromJsValue(new RangeError(getString({})))", m_js)
-            }
-            Expr::ReferenceErrorNew(m) => {
-                let m_js = self.emit_js_expr(m, locals);
-                format!("fromJsValue(new ReferenceError(getString({})))", m_js)
-            }
-            Expr::SyntaxErrorNew(m) => {
-                let m_js = self.emit_js_expr(m, locals);
-                format!("fromJsValue(new SyntaxError(getString({})))", m_js)
-            }
-            Expr::AggregateErrorNew { errors, message } => {
-                let e_js = self.emit_js_expr(errors, locals);
-                let m_js = self.emit_js_expr(message, locals);
-                format!(
-                    "fromJsValue(new AggregateError(toJsValue({}), getString({})))",
-                    e_js, m_js
-                )
-            }
-            Expr::JsonParse(val) => {
-                let v = self.emit_js_expr(val, locals);
-                format!("fromJsValue(JSON.parse(getString({})))", v)
-            }
-            Expr::JsonStringify(val) => {
-                let v = self.emit_js_expr(val, locals);
-                format!("fromJsValue(JSON.stringify(toJsValue({})))", v)
-            }
-            Expr::This => "__this".to_string(),
-            Expr::IndexGet { object, index } => {
-                let obj = self.emit_js_expr(object, locals);
-                let idx = self.emit_js_expr(index, locals);
-                format!("fromJsValue(toJsValue({})[toJsValue({})])", obj, idx)
-            }
-            Expr::IndexSet {
-                object,
-                index,
-                value,
-            } => {
-                let obj = self.emit_js_expr(object, locals);
-                let idx = self.emit_js_expr(index, locals);
-                let val = self.emit_js_expr(value, locals);
-                format!(
-                    "(toJsValue({})[toJsValue({})] = toJsValue({}), {})",
-                    obj, idx, val, val
-                )
-            }
-            Expr::ArrayPush { array_id, value } => {
-                let arr = locals
-                    .get(array_id)
-                    .cloned()
-                    .unwrap_or_else(|| format!("__l{}", array_id));
-                let val = self.emit_js_expr(value, locals);
-                format!("fromJsValue(toJsValue({}).push(toJsValue({})))", arr, val)
-            }
-            Expr::StringCoerce(val) => {
-                let v = self.emit_js_expr(val, locals);
-                format!("fromJsValue(String(toJsValue({})))", v)
-            }
-            Expr::MathFloor(x) => {
-                let v = self.emit_js_expr(x, locals);
-                format!("Math.floor({})", v)
-            }
-            Expr::MathCeil(x) => {
-                let v = self.emit_js_expr(x, locals);
-                format!("Math.ceil({})", v)
-            }
-            Expr::MathRound(x) => {
-                let v = self.emit_js_expr(x, locals);
-                format!("Math.round({})", v)
-            }
-            Expr::MathAbs(x) => {
-                let v = self.emit_js_expr(x, locals);
-                format!("Math.abs({})", v)
-            }
-            Expr::MathRandom => "Math.random()".to_string(),
-            Expr::DateNow => "Date.now()".to_string(),
-            Expr::Sequence(exprs) => {
-                if exprs.is_empty() {
-                    "u64ToF64(TAG_UNDEFINED)".to_string()
-                } else {
-                    let parts: Vec<String> =
-                        exprs.iter().map(|e| self.emit_js_expr(e, locals)).collect();
-                    format!("({})", parts.join(", "))
-                }
-            }
-            Expr::ExternFuncRef { name, .. } => {
-                // Issue #1071: prefer imported-variable global over a like-named
-                // function. In JS context the WASM exports include a getter for
-                // each global as `wasmInstance.exports.__wasm_global_<idx>` (the
-                // WASM emitter exports every global by index — see ExportSection
-                // population below). Reading the global returns the i64 NaN-box
-                // bits matching the source module's let slot.
-                let mod_key = (self.current_mod_idx, name.clone());
-                if let Some(&gidx) = self.imported_var_globals.get(&mod_key) {
-                    format!(
-                        "u64ToF64(wasmInstance.exports.__wasm_global_{}.value)",
-                        gidx
-                    )
-                } else if let Some(&func_idx) = self.func_name_map.get(name) {
-                    format!("fromJsValue(wasmInstance.exports.__wasm_func_{})", func_idx)
-                } else {
-                    "u64ToF64(TAG_UNDEFINED)".to_string()
-                }
-            }
-            Expr::FuncRef(id) => {
-                if let Some(&func_idx) = self.func_map.get(id) {
-                    format!("fromJsValue(wasmInstance.exports.__wasm_func_{})", func_idx)
-                } else {
-                    "u64ToF64(TAG_UNDEFINED)".to_string()
-                }
-            }
-            Expr::New {
-                class_name, args, ..
-            } => {
-                let args_js: Vec<String> = args
-                    .iter()
-                    .map(|a| format!("toJsValue({})", self.emit_js_expr(a, locals)))
-                    .collect();
-                format!(
-                    "fromJsValue(new (toJsValue(fromJsValue('{}')))({}))",
-                    class_name,
-                    args_js.join(", ")
-                )
-            }
-            Expr::InstanceOf { expr, ty, .. } => {
-                let e = self.emit_js_expr(expr, locals);
-                // Use the bridge instanceof check
-                format!(
-                    "(toJsValue({}) instanceof {} ? u64ToF64(TAG_TRUE) : u64ToF64(TAG_FALSE))",
-                    e, ty
-                )
-            }
-            Expr::TypeOf(operand) => {
-                let o = self.emit_js_expr(operand, locals);
-                format!("fromJsValue(typeof toJsValue({}))", o)
-            }
-            Expr::Void(e) => {
-                let v = self.emit_js_expr(e, locals);
-                format!("({}, u64ToF64(TAG_UNDEFINED))", v)
-            }
-            Expr::Delete(e) => {
-                let v = self.emit_js_expr(e, locals);
-                format!("(delete toJsValue({}), u64ToF64(TAG_TRUE))", v)
-            }
-            _ => {
-                // Fallback: return undefined for unhandled expressions
-                "u64ToF64(TAG_UNDEFINED)".to_string()
-            }
-        }
-    }
-
-    fn collect_strings(&mut self, module: &perry_hir::ir::Module) {
-        // Pre-intern common strings used by bridge calls
-        self.intern_string("Authorization");
-        self.intern_string("POST");
-        self.intern_string("GET");
-        self.intern_string("");
-
-        // Pre-intern ALL bridge function names for mem_call dispatch
-        let bridge_names = [
-            "console_log",
-            "console_warn",
-            "console_error",
-            "string_concat",
-            "js_add",
-            "string_eq",
-            "string_len",
-            "jsvalue_to_string",
-            "is_truthy",
-            "js_strict_eq",
-            "math_floor",
-            "math_ceil",
-            "math_round",
-            "math_abs",
-            "math_sqrt",
-            "math_pow",
-            "math_random",
-            "math_log",
-            "date_now",
-            "js_typeof",
-            "math_min",
-            "math_max",
-            "parse_int",
-            "parse_float",
-            "js_mod",
-            "is_null_or_undefined",
-            "object_new",
-            "object_set",
-            "object_get",
-            "object_get_dynamic",
-            "object_set_dynamic",
-            "object_delete",
-            "object_delete_dynamic",
-            "object_keys",
-            "object_values",
-            "object_entries",
-            "object_has_property",
-            "object_assign",
-            "array_new",
-            "array_push",
-            "array_pop",
-            "array_get",
-            "array_set",
-            "array_length",
-            "array_slice",
-            "array_splice",
-            "array_shift",
-            "array_unshift",
-            "array_join",
-            "array_index_of",
-            "array_includes",
-            "array_concat",
-            "array_reverse",
-            "array_flat",
-            "array_is_array",
-            "array_from",
-            "array_push_spread",
-            "string_charAt",
-            "string_substring",
-            "string_indexOf",
-            "string_slice",
-            "string_toLowerCase",
-            "string_toUpperCase",
-            "string_trim",
-            "string_includes",
-            "string_startsWith",
-            "string_endsWith",
-            "string_replace",
-            "string_split",
-            "string_fromCharCode",
-            "string_padStart",
-            "string_padEnd",
-            "string_repeat",
-            "string_match",
-            "math_log2",
-            "math_log10",
-            // Issue #133 item 4: trig / exp / sign / trunc / cbrt / hypot etc.
-            "math_sin",
-            "math_cos",
-            "math_tan",
-            "math_asin",
-            "math_acos",
-            "math_atan",
-            "math_atan2",
-            "math_sinh",
-            "math_cosh",
-            "math_tanh",
-            "math_asinh",
-            "math_acosh",
-            "math_atanh",
-            "math_exp",
-            "math_expm1",
-            "math_log1p",
-            "math_sign",
-            "math_trunc",
-            "math_cbrt",
-            "math_hypot",
-            "math_fround",
-            "math_clz32",
-            "closure_new",
-            "closure_set_capture",
-            "closure_call_0",
-            "closure_call_1",
-            "closure_call_2",
-            "closure_call_3",
-            "closure_call_spread",
-            "array_map",
-            "array_filter",
-            "array_forEach",
-            "array_reduce",
-            "array_find",
-            "array_find_index",
-            "array_sort",
-            "array_some",
-            "array_every",
-            "class_new",
-            "class_set_method",
-            "class_call_method",
-            "class_get_field",
-            "class_set_field",
-            "class_set_static",
-            "class_get_static",
-            "class_instanceof",
-            "json_parse",
-            "json_stringify",
-            "map_new",
-            "map_set",
-            "map_get",
-            "map_has",
-            "map_delete",
-            "map_size",
-            "map_clear",
-            "map_entries",
-            "map_keys",
-            "map_values",
-            "set_new",
-            "set_new_from_array",
-            "set_add",
-            "set_has",
-            "set_delete",
-            "set_size",
-            "set_clear",
-            "set_values",
-            "date_new_val",
-            "date_get_time",
-            "date_to_iso_string",
-            "date_get_full_year",
-            "date_get_month",
-            "date_get_date",
-            "date_get_day",
-            "date_get_hours",
-            "date_get_minutes",
-            "date_get_seconds",
-            "date_get_milliseconds",
-            "error_new",
-            "error_message",
-            "regexp_new",
-            "regexp_test",
-            "number_coerce",
-            "is_nan",
-            "is_finite",
-            "console_log_multi",
-            "class_set_parent",
-            "try_start",
-            "try_end",
-            "throw_value",
-            "has_exception",
-            "get_exception",
-            "url_parse",
-            "url_get_href",
-            "url_get_pathname",
-            "url_get_hostname",
-            "url_get_port",
-            "url_get_search",
-            "url_get_hash",
-            "url_get_origin",
-            "url_get_protocol",
-            "url_get_search_params",
-            "searchparams_get",
-            "searchparams_has",
-            "searchparams_set",
-            "searchparams_append",
-            "searchparams_delete",
-            "searchparams_to_string",
-            "crypto_random_uuid",
-            "crypto_random_bytes",
-            "path_join",
-            "path_dirname",
-            "path_basename",
-            "path_extname",
-            "path_resolve",
-            "path_is_absolute",
-            "os_platform",
-            "process_argv",
-            "process_cwd",
-            "buffer_alloc",
-            "buffer_from_string",
-            "buffer_to_string",
-            "buffer_get",
-            "buffer_set",
-            "buffer_length",
-            "buffer_slice",
-            "buffer_concat",
-            "uint8array_new",
-            "uint8array_from",
-            "uint8array_length",
-            "uint8array_get",
-            "uint8array_set",
-            "set_timeout",
-            "set_interval",
-            "clear_timeout",
-            "clear_interval",
-            "response_status",
-            "response_ok",
-            "response_headers_get",
-            "response_url",
-            "buffer_copy",
-            "buffer_write",
-            "buffer_equals",
-            "buffer_is_buffer",
-            "buffer_byte_length",
-            "crypto_sha256",
-            "crypto_md5",
-            "fetch_url",
-            "fetch_with_options",
-            "response_json",
-            "response_text",
-            "promise_new",
-            "promise_resolve",
-            "promise_then",
-            "await_promise",
-        ];
-        for name in &bridge_names {
-            self.intern_string(name);
-        }
-
-        for func in &module.functions {
-            self.collect_strings_in_stmts(&func.body);
-        }
-        for stmt in &module.init {
-            self.collect_strings_in_stmt(stmt);
-        }
-        for global in &module.globals {
-            if let Some(init) = &global.init {
-                self.collect_strings_in_expr(init);
-            }
-        }
-        // Collect enum names and member names/values
-        for enum_def in &module.enums {
-            self.intern_string(&enum_def.name);
-            for member in &enum_def.members {
-                self.intern_string(&member.name);
-                match &member.value {
-                    EnumValue::String(s) => {
-                        self.intern_string(s);
-                        self.enum_values.insert(
-                            (enum_def.name.clone(), member.name.clone()),
-                            EnumResolvedValue::String(s.clone()),
-                        );
-                    }
-                    EnumValue::Number(n) => {
-                        self.enum_values.insert(
-                            (enum_def.name.clone(), member.name.clone()),
-                            EnumResolvedValue::Number(*n as f64),
-                        );
-                    }
-                }
-            }
-        }
-        // Collect class names and method/field names
-        for class in &module.classes {
-            self.intern_string(&class.name);
-            if let Some(parent) = &class.extends_name {
-                self.intern_string(parent);
-            }
-            if let Some(ctor) = &class.constructor {
-                self.collect_strings_in_stmts(&ctor.body);
-                for param in &ctor.params {
-                    if let Some(default) = &param.default {
-                        self.collect_strings_in_expr(default);
-                    }
-                }
-            }
-            for method in &class.methods {
-                self.intern_string(&method.name);
-                self.collect_strings_in_stmts(&method.body);
-            }
-            for method in &class.static_methods {
-                self.intern_string(&method.name);
-                self.collect_strings_in_stmts(&method.body);
-            }
-            for (name, getter) in &class.getters {
-                self.intern_string(name);
-                self.intern_string(&format!("__get_{}", name));
-                self.collect_strings_in_stmts(&getter.body);
-            }
-            for (name, setter) in &class.setters {
-                self.intern_string(name);
-                self.intern_string(&format!("__set_{}", name));
-                self.collect_strings_in_stmts(&setter.body);
-            }
-            for field in &class.fields {
-                self.intern_string(&field.name);
-                if let Some(init) = &field.init {
-                    self.collect_strings_in_expr(init);
-                }
-            }
-            for field in &class.static_fields {
-                self.intern_string(&field.name);
-                if let Some(init) = &field.init {
-                    self.collect_strings_in_expr(init);
-                }
-            }
-        }
-    }
-
-    fn collect_strings_in_stmts(&mut self, stmts: &[Stmt]) {
-        for stmt in stmts {
-            self.collect_strings_in_stmt(stmt);
-        }
-    }
-
-    fn collect_strings_in_stmt(&mut self, stmt: &Stmt) {
-        match stmt {
-            Stmt::Let { init, .. } => {
-                if let Some(e) = init {
-                    self.collect_strings_in_expr(e);
-                }
-            }
-            Stmt::Expr(e) => self.collect_strings_in_expr(e),
-            Stmt::Return(e) => {
-                if let Some(e) = e {
-                    self.collect_strings_in_expr(e);
-                }
-            }
-            Stmt::If {
-                condition,
-                then_branch,
-                else_branch,
-            } => {
-                self.collect_strings_in_expr(condition);
-                self.collect_strings_in_stmts(then_branch);
-                if let Some(eb) = else_branch {
-                    self.collect_strings_in_stmts(eb);
-                }
-            }
-            Stmt::While { condition, body } => {
-                self.collect_strings_in_expr(condition);
-                self.collect_strings_in_stmts(body);
-            }
-            Stmt::DoWhile { body, condition } => {
-                self.collect_strings_in_stmts(body);
-                self.collect_strings_in_expr(condition);
-            }
-            Stmt::Labeled { body, .. } => {
-                self.collect_strings_in_stmt(body);
-            }
-            Stmt::For {
-                init,
-                condition,
-                update,
-                body,
-            } => {
-                if let Some(i) = init {
-                    self.collect_strings_in_stmt(i);
-                }
-                if let Some(c) = condition {
-                    self.collect_strings_in_expr(c);
-                }
-                if let Some(u) = update {
-                    self.collect_strings_in_expr(u);
-                }
-                self.collect_strings_in_stmts(body);
-            }
-            Stmt::Throw(e) => self.collect_strings_in_expr(e),
-            Stmt::Try {
-                body,
-                catch,
-                finally,
-            } => {
-                self.collect_strings_in_stmts(body);
-                if let Some(c) = catch {
-                    self.collect_strings_in_stmts(&c.body);
-                }
-                if let Some(f) = finally {
-                    self.collect_strings_in_stmts(f);
-                }
-            }
-            Stmt::Switch {
-                discriminant,
-                cases,
-            } => {
-                self.collect_strings_in_expr(discriminant);
-                for case in cases {
-                    if let Some(t) = &case.test {
-                        self.collect_strings_in_expr(t);
-                    }
-                    self.collect_strings_in_stmts(&case.body);
-                }
-            }
-            Stmt::Break | Stmt::Continue | Stmt::LabeledBreak(_) | Stmt::LabeledContinue(_) => {}
-            Stmt::PreallocateBoxes(_) => {}
-        }
-    }
-
-    fn collect_strings_in_expr(&mut self, expr: &Expr) {
-        match expr {
-            Expr::String(s) => {
-                self.intern_string(s);
-            }
-            Expr::Binary { left, right, .. }
-            | Expr::Compare { left, right, .. }
-            | Expr::Logical { left, right, .. } => {
-                self.collect_strings_in_expr(left);
-                self.collect_strings_in_expr(right);
-            }
-            Expr::Unary { operand, .. } => self.collect_strings_in_expr(operand),
-            Expr::Call { callee, args, .. } => {
-                self.collect_strings_in_expr(callee);
-                for a in args {
-                    self.collect_strings_in_expr(a);
-                }
-            }
-            Expr::LocalSet(_, val) | Expr::GlobalSet(_, val) => {
-                self.collect_strings_in_expr(val);
-            }
-            Expr::Conditional {
-                condition,
-                then_expr,
-                else_expr,
-            } => {
-                self.collect_strings_in_expr(condition);
-                self.collect_strings_in_expr(then_expr);
-                self.collect_strings_in_expr(else_expr);
-            }
-            Expr::Closure { body, .. } => {
-                self.collect_strings_in_stmts(body);
-            }
-            Expr::NativeMethodCall {
-                module,
-                method,
-                args,
-                class_name,
-                object,
-            } => {
-                for a in args {
-                    self.collect_strings_in_expr(a);
-                }
-                if let Some(obj) = object {
-                    self.collect_strings_in_expr(obj);
-                }
-                // Pre-intern bridge name for UI calls
-                let normalized = module.strip_prefix("node:").unwrap_or(module);
-                if normalized == "perry/ui" || normalized == "perry/system" {
-                    let bridge_name = map_ui_method(method, class_name.as_deref());
-                    self.intern_string(bridge_name);
-                }
-                if normalized == "perry/thread" {
-                    match method.as_str() {
-                        "parallelMap" => {
-                            self.intern_string("thread_parallel_map");
-                        }
-                        "parallelFilter" => {
-                            self.intern_string("thread_parallel_filter");
-                        }
-                        "spawn" => {
-                            self.intern_string("thread_spawn");
-                        }
-                        _ => {}
-                    }
-                }
-            }
-            Expr::Array(elems) => {
-                for e in elems {
-                    self.collect_strings_in_expr(e);
-                }
-            }
-            Expr::Object(fields) => {
-                for (k, v) in fields {
-                    self.intern_string(k);
-                    self.collect_strings_in_expr(v);
-                }
-            }
-            Expr::PropertyGet { object, property } => {
-                self.collect_strings_in_expr(object);
-                self.intern_string(property);
-            }
-            Expr::PropertySet {
-                object,
-                value,
-                property,
-                ..
-            } => {
-                self.collect_strings_in_expr(object);
-                self.collect_strings_in_expr(value);
-                self.intern_string(property);
-            }
-            Expr::IndexGet { object, index } => {
-                self.collect_strings_in_expr(object);
-                self.collect_strings_in_expr(index);
-            }
-            Expr::IndexSet {
-                object,
-                index,
-                value,
-            } => {
-                self.collect_strings_in_expr(object);
-                self.collect_strings_in_expr(index);
-                self.collect_strings_in_expr(value);
-            }
-            Expr::Await(e) | Expr::TypeOf(e) | Expr::Void(e) => {
-                self.collect_strings_in_expr(e);
-            }
-            Expr::New { args, .. } => {
-                for a in args {
-                    self.collect_strings_in_expr(a);
-                }
-            }
-            Expr::Update { .. } => {}
-            Expr::Sequence(exprs) => {
-                for e in exprs {
-                    self.collect_strings_in_expr(e);
-                }
-            }
-            Expr::EnumMember {
-                enum_name,
-                member_name,
-            } => {
-                self.intern_string(enum_name);
-                self.intern_string(member_name);
-            }
-            Expr::StaticFieldGet {
-                class_name,
-                field_name,
-            }
-            | Expr::StaticFieldSet {
-                class_name,
-                field_name,
-                ..
-            } => {
-                self.intern_string(class_name);
-                self.intern_string(field_name);
-            }
-            Expr::StaticMethodCall {
-                class_name,
-                method_name,
-                args,
-            } => {
-                self.intern_string(class_name);
-                self.intern_string(method_name);
-                for a in args {
-                    self.collect_strings_in_expr(a);
-                }
-            }
-            Expr::InstanceOf { expr, ty, .. } => {
-                self.collect_strings_in_expr(expr);
-                self.intern_string(ty);
-            }
-            Expr::In { property, object } => {
-                self.collect_strings_in_expr(property);
-                self.collect_strings_in_expr(object);
-            }
-            Expr::Delete(e) => self.collect_strings_in_expr(e),
-            Expr::RegExp { pattern, flags } => {
-                self.intern_string(pattern);
-                self.intern_string(flags);
-            }
-            Expr::RegExpTest { regex, string } => {
-                self.collect_strings_in_expr(regex);
-                self.collect_strings_in_expr(string);
-            }
-            Expr::StringMatch { string, regex } => {
-                self.collect_strings_in_expr(string);
-                self.collect_strings_in_expr(regex);
-            }
-            Expr::StringReplace {
-                string,
-                pattern,
-                replacement,
-            } => {
-                self.collect_strings_in_expr(string);
-                self.collect_strings_in_expr(pattern);
-                self.collect_strings_in_expr(replacement);
-            }
-            Expr::StringSplit(a, b) => {
-                self.collect_strings_in_expr(a);
-                self.collect_strings_in_expr(b);
-            }
-            Expr::StringFromCharCode(e) | Expr::StringFromCodePoint(e) | Expr::StringCoerce(e) => {
-                self.collect_strings_in_expr(e);
-            }
-            Expr::StringAt { string, index } | Expr::StringCodePointAt { string, index } => {
-                self.collect_strings_in_expr(string);
-                self.collect_strings_in_expr(index);
-            }
-            Expr::ObjectSpread { parts } => {
-                for (key_opt, val) in parts {
-                    if let Some(k) = key_opt {
-                        self.intern_string(k);
-                    }
-                    self.collect_strings_in_expr(val);
-                }
-            }
-            Expr::ArraySpread(elements) => {
-                for elem in elements {
-                    match elem {
-                        ArrayElement::Expr(e) | ArrayElement::Spread(e) => {
-                            self.collect_strings_in_expr(e);
-                        }
-                    }
-                }
-            }
-            Expr::ObjectKeys(e) | Expr::ObjectValues(e) | Expr::ObjectEntries(e) => {
-                self.collect_strings_in_expr(e);
-            }
-            Expr::ObjectRest {
-                object,
-                exclude_keys,
-            } => {
-                self.collect_strings_in_expr(object);
-                for k in exclude_keys {
-                    self.intern_string(k);
-                }
-            }
-            Expr::ArrayPush { value, .. } | Expr::ArrayUnshift { value, .. } => {
-                self.collect_strings_in_expr(value);
-            }
-            Expr::ArrayPushSpread { source, .. } => {
-                self.collect_strings_in_expr(source);
-            }
-            Expr::ArraySlice { array, start, end } => {
-                self.collect_strings_in_expr(array);
-                self.collect_strings_in_expr(start);
-                if let Some(e) = end {
-                    self.collect_strings_in_expr(e);
-                }
-            }
-            Expr::ArraySplice {
-                start,
-                delete_count,
-                items,
-                ..
-            } => {
-                self.collect_strings_in_expr(start);
-                if let Some(dc) = delete_count {
-                    self.collect_strings_in_expr(dc);
-                }
-                for item in items {
-                    self.collect_strings_in_expr(item);
-                }
-            }
-            Expr::ArrayJoin { array, separator } => {
-                self.collect_strings_in_expr(array);
-                if let Some(s) = separator {
-                    self.collect_strings_in_expr(s);
-                }
-                self.intern_string(","); // default separator
-            }
-            Expr::ArrayIndexOf { array, value } | Expr::ArrayIncludes { array, value } => {
-                self.collect_strings_in_expr(array);
-                self.collect_strings_in_expr(value);
-            }
-            Expr::ArrayMap { array, callback }
-            | Expr::ArrayFilter { array, callback }
-            | Expr::ArrayForEach { array, callback }
-            | Expr::ArrayFind { array, callback }
-            | Expr::ArrayFindIndex { array, callback }
-            | Expr::ArraySort {
-                array,
-                comparator: callback,
-            } => {
-                self.collect_strings_in_expr(array);
-                self.collect_strings_in_expr(callback);
-            }
-            Expr::ArrayReduce {
-                array,
-                callback,
-                initial,
-            }
-            | Expr::ArrayReduceRight {
-                array,
-                callback,
-                initial,
-            } => {
-                self.collect_strings_in_expr(array);
-                self.collect_strings_in_expr(callback);
-                if let Some(i) = initial {
-                    self.collect_strings_in_expr(i);
-                }
-            }
-            Expr::ArrayToSorted { array, comparator } => {
-                self.collect_strings_in_expr(array);
-                if let Some(c) = comparator {
-                    self.collect_strings_in_expr(c);
-                }
-            }
-            Expr::ArrayToSpliced {
-                array,
-                start,
-                delete_count,
-                items,
-            } => {
-                self.collect_strings_in_expr(array);
-                self.collect_strings_in_expr(start);
-                self.collect_strings_in_expr(delete_count);
-                for item in items {
-                    self.collect_strings_in_expr(item);
-                }
-            }
-            Expr::ArrayWith {
-                array,
-                index,
-                value,
-            } => {
-                self.collect_strings_in_expr(array);
-                self.collect_strings_in_expr(index);
-                self.collect_strings_in_expr(value);
-            }
-            Expr::ArrayCopyWithin {
-                target, start, end, ..
-            } => {
-                self.collect_strings_in_expr(target);
-                self.collect_strings_in_expr(start);
-                if let Some(e) = end {
-                    self.collect_strings_in_expr(e);
-                }
-            }
-            Expr::ArrayFlat { array }
-            | Expr::ArrayIsArray(array)
-            | Expr::ArrayFrom(array)
-            | Expr::ArrayToReversed { array } => {
-                self.collect_strings_in_expr(array);
-            }
-            Expr::ArrayEntries(array) | Expr::ArrayKeys(array) | Expr::ArrayValues(array) => {
-                self.collect_strings_in_expr(array);
-            }
-            Expr::ArrayFromMapped { iterable, map_fn } => {
-                self.collect_strings_in_expr(iterable);
-                self.collect_strings_in_expr(map_fn);
-            }
-            Expr::MapSet { map, key, value } => {
-                self.collect_strings_in_expr(map);
-                self.collect_strings_in_expr(key);
-                self.collect_strings_in_expr(value);
-            }
-            Expr::MapGet { map, key }
-            | Expr::MapHas { map, key }
-            | Expr::MapDelete { map, key } => {
-                self.collect_strings_in_expr(map);
-                self.collect_strings_in_expr(key);
-            }
-            Expr::MapSize(e)
-            | Expr::MapClear(e)
-            | Expr::MapEntries(e)
-            | Expr::MapKeys(e)
-            | Expr::MapValues(e)
-            | Expr::MapNewFromArray(e) => {
-                self.collect_strings_in_expr(e);
-            }
-            Expr::SetNewFromArray(e)
-            | Expr::SetSize(e)
-            | Expr::SetClear(e)
-            | Expr::SetValues(e) => {
-                self.collect_strings_in_expr(e);
-            }
-            Expr::SetAdd { value, .. } => {
-                self.collect_strings_in_expr(value);
-            }
-            Expr::SetHas { set, value } | Expr::SetDelete { set, value } => {
-                self.collect_strings_in_expr(set);
-                self.collect_strings_in_expr(value);
-            }
-            Expr::DateNew(args) => {
-                for a in args {
-                    self.collect_strings_in_expr(a);
-                }
-            }
-            Expr::DateGetTime(e)
-            | Expr::DateToISOString(e)
-            | Expr::DateGetFullYear(e)
-            | Expr::DateGetMonth(e)
-            | Expr::DateGetDate(e)
-            | Expr::DateGetDay(e)
-            | Expr::DateGetHours(e)
-            | Expr::DateGetMinutes(e)
-            | Expr::DateGetSeconds(e)
-            | Expr::DateGetMilliseconds(e)
-            | Expr::DateGetUtcDay(e)
-            | Expr::DateGetUtcFullYear(e)
-            | Expr::DateGetUtcMonth(e)
-            | Expr::DateGetUtcDate(e)
-            | Expr::DateGetUtcHours(e)
-            | Expr::DateGetUtcMinutes(e)
-            | Expr::DateGetUtcSeconds(e)
-            | Expr::DateGetUtcMilliseconds(e)
-            | Expr::DateValueOf(e)
-            | Expr::DateToDateString(e)
-            | Expr::DateToTimeString(e)
-            | Expr::DateToLocaleDateString(e)
-            | Expr::DateToLocaleTimeString(e)
-            | Expr::DateToLocaleString(e)
-            | Expr::DateGetTimezoneOffset(e)
-            | Expr::DateToJSON(e)
-            | Expr::DateParse(e) => {
-                self.collect_strings_in_expr(e);
-            }
-            Expr::DateUtc(args) => {
-                for a in args {
-                    self.collect_strings_in_expr(a);
-                }
-            }
-            Expr::DateSetUtcFullYear { date, value }
-            | Expr::DateSetUtcMonth { date, value }
-            | Expr::DateSetUtcDate { date, value }
-            | Expr::DateSetUtcHours { date, value }
-            | Expr::DateSetUtcMinutes { date, value }
-            | Expr::DateSetUtcSeconds { date, value }
-            | Expr::DateSetUtcMilliseconds { date, value } => {
-                self.collect_strings_in_expr(date);
-                self.collect_strings_in_expr(value);
-            }
-            Expr::ErrorNew(msg) => {
-                if let Some(m) = msg {
-                    self.collect_strings_in_expr(m);
-                }
-            }
-            Expr::ErrorMessage(e) => {
-                self.collect_strings_in_expr(e);
-            }
-            Expr::ErrorNewWithCause { message, cause } => {
-                self.collect_strings_in_expr(message);
-                self.collect_strings_in_expr(cause);
-            }
-            Expr::TypeErrorNew(m)
-            | Expr::RangeErrorNew(m)
-            | Expr::ReferenceErrorNew(m)
-            | Expr::SyntaxErrorNew(m) => {
-                self.collect_strings_in_expr(m);
-            }
-            Expr::AggregateErrorNew { errors, message } => {
-                self.collect_strings_in_expr(errors);
-                self.collect_strings_in_expr(message);
-            }
-            Expr::JsonParse(e) | Expr::JsonStringify(e) => {
-                self.collect_strings_in_expr(e);
-            }
-            Expr::NumberCoerce(e)
-            | Expr::IsNaN(e)
-            | Expr::IsUndefinedOrBareNan(e)
-            | Expr::IsFinite(e)
-            | Expr::BigIntCoerce(e) => {
-                self.collect_strings_in_expr(e);
-            }
-            Expr::ParseInt { string, radix } => {
-                self.collect_strings_in_expr(string);
-                if let Some(r) = radix {
-                    self.collect_strings_in_expr(r);
-                }
-            }
-            Expr::ParseFloat(e) => {
-                self.collect_strings_in_expr(e);
-            }
-            Expr::PropertyUpdate {
-                object, property, ..
-            } => {
-                self.collect_strings_in_expr(object);
-                self.intern_string(property);
-            }
-            Expr::IndexUpdate { object, index, .. } => {
-                self.collect_strings_in_expr(object);
-                self.collect_strings_in_expr(index);
-            }
-            Expr::SuperCall(args) => {
-                for a in args {
-                    self.collect_strings_in_expr(a);
-                }
-            }
-            Expr::SuperMethodCall { method, args } => {
-                self.intern_string(method);
-                for a in args {
-                    self.collect_strings_in_expr(a);
-                }
-            }
-            Expr::NewDynamic { callee, args } => {
-                self.collect_strings_in_expr(callee);
-                for a in args {
-                    self.collect_strings_in_expr(a);
-                }
-            }
-            Expr::FetchWithOptions {
-                url,
-                method,
-                body,
-                headers,
-            } => {
-                self.collect_strings_in_expr(url);
-                self.collect_strings_in_expr(method);
-                self.collect_strings_in_expr(body);
-                for (key, val) in headers {
-                    self.intern_string(key);
-                    self.collect_strings_in_expr(val);
-                }
-            }
-            Expr::FetchGetWithAuth { url, auth_header } => {
-                self.collect_strings_in_expr(url);
-                self.collect_strings_in_expr(auth_header);
-            }
-            Expr::FetchPostWithAuth {
-                url,
-                auth_header,
-                body,
-            } => {
-                self.collect_strings_in_expr(url);
-                self.collect_strings_in_expr(auth_header);
-                self.collect_strings_in_expr(body);
-            }
-            _ => {}
         }
     }
 }
